requirements.txt
backcall==0.2.0
biopython==1.78
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
entrypoints==0.4
filelock==3.4.1
future==1.0.0
idna==3.6
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
matplotlib==3.3.4
nest-asyncio==1.6.0
numpy==1.19.5
olefile==0.46
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.1.2
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
regex==2023.8.8
requests==2.27.1
sacremoses==0.0.53
scipy==1.5.4
seaborn==0.11.2
sentencepiece==0.1.91
six==1.16.0
tokenizers==0.9.3
torch==1.7.0
tornado==6.1
tqdm==4.53.0
traitlets==4.3.3
transformers==3.5.1
typing-extensions==4.1.1
urllib3==1.26.18
wcwidth==0.2.13
zipp==3.6.0
